Restyle publication details page for more components
