+++ client/src/components/RecentSubmissions.vue
@@ -1,0 +1,11 @@
+<template>
+  <q-card flat>
+    <q-card-section>
+      <div class="text-h3">Recent Submissions</div>
+    </q-card-section>
+  </q-card>
+</template>
+
+<script setup></script>
+
+<style></style>
