+++ client/src/graphql/queries.js
@@ -185,6 +185,7 @@ export const GET_SUBMISSION = gql`
 export const GET_PUBLICATION = gql`
   query GetPublication($id: ID!) {
     publication(id: $id) {
+      id
       name
       is_publicly_visible
       users {
